Extract pane emulator helpers into pane_emulator.go
